StoreOwnerProfileManager was being added
